Refactor player, add needed mappers, dto-classes, and modules for require functionality

+++ muciplayer-data/src/main/java/com/andreich/data/mapper/AlbumEntityToAlbumMapper.kt
@@ -8,7 +8,7 @@ class AlbumEntityToAlbumMapper : EntityToModelMapper<AlbumEntity, Album> {
     override fun map(fromEntity: AlbumEntity): Album {
         return with(fromEntity) {
             Album(
-                id = id,
+                id = id.toInt(),
                 title = title,
                 cover = cover,
                 coverBig = coverBig,
